requirements.txt
flask
python-telegram-bot==13.15
python-dotenv
pandas
numpy
yfinance
requests
ta
joblib
scikit-learn==1.7.2
